refactor: rip the agent-shim; pi state comes from the extension (ADR 0011)
